pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skeleton-crew"
version = "0.1.0"
description = "Distributed JACK-centric control hub for film scoring and production"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Skeleton App Team"}
]
dependencies = [
    "pyyaml>=6.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "httpx>=0.25.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "PySide6>=6.6.0",
    "pyzmq>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]
audio = [
    "JACK-Client>=0.5.4",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
]
stt = [
    "vosk>=0.3.45",
    "openai-whisper>=20231117",
    "faster-whisper>=0.10.0",
]
tts = [
    "piper-tts>=1.2.0",
]
llm = [
    "ollama>=0.1.0",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
]
midi = [
    "python-rtmidi>=1.5.0",
    "mido>=1.3.0",
]
osc = [
    "python-osc>=1.8.0",
]

[project.scripts]
skeleton = "skeleton_app.cli:main"
skeleton-daemon = "skeleton_app.daemon:main"
skeleton-gui = "skeleton_app.gui.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
